% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypoimg_generic.R
\docType{data}
\name{clr5}
\alias{clr5}
\title{five colors}
\format{
An object of class \code{character} of length 5.
}
\usage{
clr5
}
\description{
\code{clr5} is a combination of five colors I like.
}
\details{
This is just a shortcut for RColorBrewer::brewer.pal(5,'Set1').
}
\examples{
ggplot(tibble( x = 1:5))+
  coord_equal()+
  geom_tile(aes(x = x, y = 1, fill = factor(x)))+
  scale_fill_manual(values = clr5)+
  theme_void()+
  theme(legend.position = 'none')

}
\keyword{datasets}
