% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypoimg_generic.R
\name{clr_lighten}
\alias{clr_lighten}
\title{Programmatically lighten colors}
\usage{
clr_lighten(color, factor = 0.2)
}
\arguments{
\item{color}{string (color, mandatory)}

\item{factor}{numeric scalar (optional), between 0 and 1, amount of lightening}
}
\description{
\code{clr_lighten} produces a lighter version on given colors.
}
\details{
superseded: Please use \code{prismatic::clr_lighten()} instead

Takes a vector of colors and lightens them by a given factor.
1 returns the input color, 0 returns white
}
\examples{
ggplot(tibble( x = 1:5))+
  coord_equal()+
  geom_tile(aes(x = x, y = 1, fill = factor(x)))+
  scale_fill_manual(values = map_chr((1:5)/5,.f = clr_lighten, color = clr2[[1]]))+
  theme_void()+
  theme(legend.position = 'none')

}
\seealso{
\code{\link{clr_darken}}
}
