% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypoimg_geom_grob.R
\name{geom_hypo_grob}
\alias{geom_hypo_grob}
\title{Provides the grob geom}
\usage{
geom_hypo_grob(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = FALSE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by aes()}

\item{data}{The data to be displayed in this layer. This shoul be a tibble
containing the grobs as one column, x and y values as columns and potential
additional aesthetics (width, height and angle). Note, that x and y should
be within 0 and 1 and that width and height represent the ration comprared
to default size of the grob.}

\item{stat}{The statistical transformation to use on the data for this layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function.}

\item{na.rm}{logical.}

\item{show.legend}{logical.}

\item{inherit.aes}{logical.}

\item{...}{parameters passed to layer}
}
\description{
\code{geom_hypo_grob} provides a geom of grobs
}
\details{
The function \code{geom_hypo_grob} provides a way to add grob annotations
as geom so that the can be distributed over differnt facets.

This geom builds heavily on the answer by baptiste on the
tidiverse github forum:
https://github.com/tidyverse/ggplot2/issues/1399

Only the parameters height, width and angle have been added.
}
