% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypoimg_flags.R
\name{hypo_anno_flag}
\alias{hypo_anno_flag}
\title{Add a flag to a ggplot}
\usage{
hypo_anno_flag(geo, xmin = -Inf, xmax = Inf, ymin = -Inf, ymax = Inf,
  ...)
}
\arguments{
\item{geo}{string skalar (manatory), one of the available flag IDs}

\item{xmin}{numeric skalar (optional), left boundary of the annotation}

\item{xmax}{numeric skalar (optional), right boundary of the annotation}

\item{ymin}{numeric skalar (optional), lower boundary of the annotation}

\item{ymax}{numeric skalar (optional), upper boundary of the annotation}
}
\description{
\code{hypo_anno_flag} adds a left facing hamlet annotation to a ggplot.
}
\details{
Hypogen comes with a set of illustrations of flags of the countries adjacent
to the Caribbean and the Golf of Mexico.
The function \code{hypo_anno_flag} uses the \code{ggplot2::annotation_custom()}
function to add a single flag to an existring ggplot.
}
\examples{
ggplot(tibble(x = 1, y = 1), aes(x = x, y = y))+
  geom_point()+
  hypo_anno_flag('mexico', xmax = 1.2, ymax = 1.2)

}
\seealso{
\code{\link{hypo_flag_palette}},
  \code{\link{hypo_anno_l}},
  \code{\link{hypo_anno_r}}
}
