% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypoimg_flags.R
\name{hypo_anno_flag_pair}
\alias{hypo_anno_flag_pair}
\title{One paired flag legend element}
\usage{
hypo_anno_flag_pair(left, right, flag_lwd = 1, flag_line_color = NA,
  flag_radius_scale = 1, circle_color = NA, circle_fill = "white",
  circle_lwd = 0.5, plot_names = FALSE, plot_name_size = 3,
  font_family = "sans", ...)
}
\arguments{
\item{left}{string skalar (manatory), one of the available flags}

\item{right}{string skalar (manatory), one of the available flags}

\item{flag_lwd}{numeric skalar (optional), line width of the optional
frame of the flag annotation}

\item{flag_line_color}{color skalar (a string, optional), line width of the optional
frame of the flag annotation}

\item{flag_radius_scale}{numeric skalar (optional), scaling of the optional
flag annotation frames diameter}

\item{circle_color}{color skalar (a string, optional), the color of the
background circle outline}

\item{circle_fill}{color skalar (a string, optional), the fill of the
background circle}

\item{circle_lwd}{numeric skalar (optional), the width of the background
circle outline}

\item{plot_names}{logical skalar (optional), should the species labels
be added?}

\item{plot_name_size}{numeric skalar (optional), the species label size}

\item{font_family}{string skalar (optional), the species label font family}
}
\description{
\code{hypo_anno_flag_pair} combines a paired flag with a colored circle
}
\details{
The function \code{hypo_anno_flag_pair} provides the basic building
block of the paired flag color legend. It combines a paired flag
annotaion with a background circle which can be coloured.

Aditionally, the flag labels can be added to the  plot.
}
\examples{
hypo_anno_flag_pair(left= 'belize', right = 'panama',
  flag_lwd = 1, flag_line_color = 'black',
  circle_color = 'black', plot_names = TRUE,
  plot_name_size = 5)

}
\seealso{
\code{\link{hypo_anno_flag_single}},
  \code{\link{hypo_anno_pair}}
}
