% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypoimg_basics.R
\name{hypo_anno_pair}
\alias{hypo_anno_pair}
\title{One paired hamlet legend element}
\usage{
hypo_anno_pair(left, right, circle_color = NA, circle_fill = "white",
  circle_lwd = 0.5, plot_names = FALSE, plot_name_size = 3,
  font_family = "sans", ...)
}
\arguments{
\item{left}{string skalar (manatory), one of the available species}

\item{right}{string skalar (manatory), one of the available species}

\item{circle_color}{color skalar (a string, optional), the color of the
background circle outline}

\item{circle_fill}{color skalar (a string, optional), the fill of the
background circle}

\item{circle_lwd}{numeric skalar (optional), the width of the background
circle outline}

\item{plot_names}{logical skalar (optional), should the species labels
be added?}

\item{plot_name_size}{numeric skalar (optional), the species label size}

\item{font_family}{string skalar (optional), the species label font family}
}
\description{
\code{hypo_anno_pair} combines a paired hamlet with a colored circle
}
\details{
The function \code{hypo_anno_pair} provides the basic building block
of the paired hamlet color legend. It combines a paired hamlet annotaion
with a background circle which can be coloured.

Aditionally, the species labels can be added to the  plot.

Available species:
  "aberrans", "atlahua", "castroaguirrei", "chlorurus", "ecosur",
  "floridae", "gemma", "gummigutta", "guttavarius", "indigo",
  "liberte", "maculiferus", "maya", "nigricans", "providencianus",
  "puella", "randallorum", "tan", "unicolor"
}
\examples{
hypo_anno_pair('indigo','puella',
  circle_color = 'black',plot_names = TRUE)

}
\seealso{
\code{\link{hypo_anno_single}},
  \code{\link{hypo_anno_flag_pair}}
}
