% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypoimg_basics.R
\name{hypo_anno_pair_split}
\alias{hypo_anno_pair_split}
\title{One paired hamlet legend element with two fills}
\usage{
hypo_anno_pair_split(left, right, circle_color = NA,
  circle_fill_left = "white", circle_fill_right = "lightgray",
  circle_lwd = 0.5, plot_names = FALSE, plot_name_size = 3,
  font_family = "sans", ...)
}
\arguments{
\item{left}{string skalar (manatory), one of the available species}

\item{right}{string skalar (manatory), one of the available species}

\item{circle_color}{color skalar (a string, optional), the color of the
background circle outline}

\item{circle_fill_left}{color skalar (a string, optional), the fill of the
left background}

\item{circle_fill_right}{color skalar (a string, optional), the fill of the
right background}

\item{circle_lwd}{numeric skalar (optional), the width of the background
circle outline}

\item{plot_names}{logical skalar (optional), should the species labels
be added?}

\item{plot_name_size}{numeric skalar (optional), the species label size}

\item{font_family}{string skalar (optional), the species label font family}
}
\description{
\code{hypo_anno_pair_split} combines a paired hamlet with a two-colored circle
}
\details{
The function \code{hypo_anno_pair_split} provides the basic building block
of the paired hamlet color legend. It combines a paired hamlet annotaion
with a background circle which can be coloured using two fills for
the two species.

Aditionally, the species labels can be added to the  plot.

Available species:
  "aberrans", "atlahua", "castroaguirrei", "chlorurus", "ecosur",
  "floridae", "gemma", "gumigutta", "guttavarius", "indigo",
  "liberte", "maculiferus", "maya", "nigricans", "providencianus",
  "puella", "randallorum", "tan", "unicolor"
}
\examples{
hypo_anno_pair_split(left = 'puella', right = 'indigo',
  circle_color = 'black', plot_names = TRUE)

}
\seealso{
\code{\link{hypo_anno_single}},
  \code{\link{hypo_anno_flag_pair}}
}
