% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypoimg_basics.R
\name{hypo_anno_r}
\alias{hypo_anno_r}
\title{Add a right facing hamlet to a ggplot}
\usage{
hypo_anno_r(species, xmin = -Inf, xmax = Inf, ymin = -Inf, ymax = Inf,
  ...)
}
\arguments{
\item{species}{string skalar (manatory), one of "aberrans","atlahua",
"castroaguirrei","chlorurus","ecosur","floridae","gemma","gumigutta",
"guttavarius","indigo","liberte","maculiferus","maya","nigricans",
"providencianus","puella","randallorum","tan","unicolor"}

\item{xmin}{numeric skalar (optional), left boundary of the annotation}

\item{xmax}{numeric skalar (optional), right boundary of the annotation}

\item{ymin}{numeric skalar (optional), lower boundary of the annotation}

\item{ymax}{numeric skalar (optional), upper boundary of the annotation}
}
\description{
\code{hypo_anno_r} adds a right facing hamlet annotation to a ggplot.
}
\details{
Hypogen comes with a set of illustrations of differnt halmet species.
The function \code{hypo_anno_r} uses the \code{ggplot2::annotation_custom()}
function to add a single right facing hamlet to a existring ggplot.
}
\examples{
ggplot(tibble(x = 1, y = 1),
aes(x = x, y = y))+
  geom_point()+
  hypo_anno_l('unicolor', xmax = 1.2, ymax = 1.2)+
  hypo_anno_r('aberrans', xmin = .8, ymin = .8)

}
\seealso{
\code{\link{hypo_anno_l}},
  \code{\link{hypo_anno_flag}}
}
