% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypoimg_basics.R
\name{hypo_anno_single}
\alias{hypo_anno_single}
\title{One single hamlet legend element}
\usage{
hypo_anno_single(species, circle_color = NA, circle_fill = "white",
  circle_lwd = 0.5, plot_names = FALSE, plot_name_size = 3,
  font_family = "sans", xlims = 0.45, ylims = 0.335, ...)
}
\arguments{
\item{species}{string skalar (manatory), one of the available species}

\item{circle_color}{color skalar (a string, optional), the color of the
background circle outline}

\item{circle_fill}{color skalar (a string, optional), the fill of the
background circle}

\item{circle_lwd}{numeric skalar (optional), the width of the background
circle outline}

\item{plot_names}{logical skalar (optional), should the species label
be added?}

\item{plot_name_size}{numeric skalar (optional), the species label size}

\item{font_family}{string skalar (optional), the species label font family}

\item{xlims}{numeric skalar (optional), modifies the horizontal extent of the hamlet}
}
\description{
\code{hypo_anno_single} combines a single hamlet with a colored circle
}
\details{
The function \code{hypo_anno_single} provides the basic building block
of the single hamlet color legend. It combines a single hamlet annotaion
with a background circle which can be coloured.

Aditionally, the species label can be added to the  plot.

Available species:
  "aberrans", "atlahua", "castroaguirrei", "chlorurus", "ecosur",
  "floridae", "gemma", "gumigutta", "guttavarius", "indigo",
  "liberte", "maculiferus", "maya", "nigricans", "providencianus",
  "puella", "randallorum", "tan", "unicolor"
}
\examples{
hypo_anno_single('indigo',circle_color = 'black',
  plot_names = TRUE)

}
\seealso{
\code{\link{hypo_anno_pair}},
  \code{\link{hypo_anno_flag_single}}
}
