% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypoimg_data.R
\docType{data}
\name{hypo_coral_img}
\alias{hypo_coral_img}
\title{Several coral drawings.}
\format{
A tibble with 6 rows and 3 variables:
\describe{
\item{side}{string, orientation of the drawing (left/ right)}
\item{type}{string, type of coral}
\item{grob}{grob ({grid} object), the actual drawing}
}
}
\source{
drawing by KH.
}
\usage{
hypo_coral_img
}
\description{
A data set holding several coral drawings.
}
\examples{
ggplot() +
  geom_hypo_grob(data = hypo_coral_img,
                 aes(x = .5, y = .5, grob = grob)) +
  facet_grid(side ~ type)
}
\keyword{datasets}
