% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypoimg_data.R
\docType{data}
\name{hypo_flag}
\alias{hypo_flag}
\title{Several Country Flags.}
\format{
A tibble with 41 rows and 2 variables:
\describe{
\item{geo}{string, name of the country}
\item{grob}{grob ({grid} object), the actual flag}
}
}
\source{
modified from \url{wikipedia.org}.
}
\usage{
hypo_flag
}
\description{
A data set holding several flags.
}
\examples{
ggplot() +
  geom_hypo_grob(data = hypo_coral_img,
                 aes(x = .5, y = .5, grob = grob)) +
  facet_grid(side ~ type)
}
\keyword{datasets}
