% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypoimg_data.R
\docType{data}
\name{hypo_generic_img}
\alias{hypo_generic_img}
\title{Generic hamlet drawing.}
\format{
A tibble with 2 rows and 2 variables:
\describe{
\item{side}{string, orientation of the drawing (left/ right)}
\item{grob}{grob ({grid} object), the actual drawing}
}
}
\source{
drawing by KH.
}
\usage{
hypo_generic_img
}
\description{
A data set holding a generic hamlet drawing.
}
\examples{
ggplot() +
  geom_hypo_grob(data = hypo_generic_img,
                 aes(x = .5, y = .5, grob = grob)) +
  facet_wrap(side ~ .)
}
\keyword{datasets}
