% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypoimg_geom_grob.R
\docType{data}
\name{hypo_geom_grob_custom}
\alias{hypo_geom_grob_custom}
\title{The custom grob geom}
\format{
An object of class \code{hypo_geom_grob_custom} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
hypo_geom_grob_custom
}
\description{
This geom builds heavily on the answer by baptiste on the
tidiverse github forum:
https://github.com/tidyverse/ggplot2/issues/1399
}
\details{
Only the parameters height, width and angle have been added.
}
\seealso{
\code{\link{geom_hypo_grob}}
}
\keyword{datasets}
