% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypoimg_generic.R
\name{hypo_hamlet_generic}
\alias{hypo_hamlet_generic}
\title{Add a generic hamlet to a ggplot}
\usage{
hypo_hamlet_generic(col = "darkgray", x_in = 0.5, y_in = 0.5,
  width_in = 1, height_in = 1, face = c("l", "r"), ...)
}
\arguments{
\item{col}{string  (color, optional), defaul = darkgray}

\item{width_in}{numeric skalar (optional), width (in graph x axis units)}

\item{height_in}{numeric skalar (optional), width (in graph y axis units)}

\item{face}{string skalar (optional), one of l,r side the hamlet is facing}

\item{x}{numeric skalar (optional), horizontal center for hamlet placement}

\item{y}{numeric skalar (optional), vertical center for hamlet placement}
}
\description{
\code{hypo_hamlet_generic} adds a generic hamlet with custom color.
}
\details{
This function allows the placement of a generic hamlet within a ggplot.
It is based on \code{ggplot2::annotation_custom()} and allows to specify
the coloration of the generic hamlet.
}
\examples{
ggplot()+
  hypo_hamlet_generic(col = clr2[[1]])

}
\seealso{
\code{\link{hypo_coral_generic}}
}
