% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypoimg_data.R
\docType{data}
\name{hypo_img}
\alias{hypo_img}
\title{Hamlet species drawings.}
\format{
A tibble with 22 rows and 4 variables:
\describe{
\item{spec}{string, species name}
\item{geno}{string, genus (sic.), abbreviation}
\item{l}{grob ({grid} object), the left facing drawings}
\item{r}{grob ({grid} object), the right facing drawings}
}
}
\source{
drawing by KH.
}
\usage{
hypo_img
}
\description{
A data set holding species specific hamlet drawings.
}
\examples{
ggplot() +
  geom_hypo_grob(data = hypo_img,
                 aes(x = .5, y = .5, grob = l)) +
  facet_wrap(spec ~ .)

ggplot() +
  geom_hypo_grob(data = hypo_img,
                 aes(x = .5, y = .5, grob = r)) +
  facet_wrap(spec ~ .)
}
\keyword{datasets}
