% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypoimg_flags.R
\name{hypo_legend_flag_pair}
\alias{hypo_legend_flag_pair}
\title{Constructs a legend of paired flags}
\usage{
hypo_legend_flag_pair(
  left,
  right,
  color_map,
  flag_lwd = 1,
  flag_line_color = NA,
  flag_radius_scale = 1,
  circle_color = NA,
  circle_lwd = 0.5,
  plot_names = FALSE,
  plot_name_size = 3,
  font_family = "sans",
  plot = TRUE
)
}
\arguments{
\item{left}{string vector (manatory), one of the available flag IDs}

\item{right}{string vector (manatory), one of the available flag IDs}

\item{color_map}{color vector (a string, optional), the color map (must
match the species vector in length)}

\item{flag_lwd}{numeric skalar (optional), line width of the optional
frame of the flag annotation}

\item{flag_line_color}{color skalar (a string, optional), line width of the
optional frame of the flag annotation}

\item{flag_radius_scale}{numeric skalar (optional), scaling of the optional
flag annotation frames diameter}

\item{circle_color}{color skalar (a string, optional), the color of the
background circle outline}

\item{circle_lwd}{numeric skalar (optional), the width of the background
circle outlines}

\item{plot_names}{logical skalar (optional), should the species label
be added?}

\item{plot_name_size}{numeric skalar (optional), the species label size}

\item{font_family}{string skalar (optional), the species label font family}

\item{plot}{logical scalar (optional), toggle the output to be either a plot
or a list of plots}
}
\description{
\code{hypo_legend_flag_pair} combines several paired flag legend elements
}
\details{
The function \code{hypo_legend_flag_pair} constructs a paired flag legend
from two vectors of flags and a choosen color map (matching in length).
}
\examples{
clr <- RColorBrewer::brewer.pal(3,'Blues')

left_flag <- c('anguilla','barbados','trinidad_and_tobago')
right_flag <- c('usa','virgin_islands','cuba')

hypo_legend_flag_pair(left = left_flag, right = right_flag,
                      color_map = clr,
                      flag_lwd = 1, flag_line_color = 'black',
                      flag_radius_scale = 1,
                      circle_color = 'black', plot_names = TRUE)
}
\seealso{
\code{\link{hypo_legend_single}},
\code{\link{hypo_legend_pair}}
}
