% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypoimg_basics.R
\name{hypo_legend_single}
\alias{hypo_legend_single}
\title{Constructs a legend of single hamlets}
\usage{
hypo_legend_single(species, color_map, circle_color = NA, circle_lwd = 0.5,
  plot_names = FALSE, plot_name_size = 3, font_family = "sans",
  plot = TRUE)
}
\arguments{
\item{species}{string vector (manatory), can only contain available species}

\item{color_map}{color vector (a string, optional), the color map (must
match the species vector in length)}

\item{circle_color}{color skalar (a string, optional), the color of the
background circle outline}

\item{circle_lwd}{numeric skalar (optional), the width of the background
circle outlines}

\item{plot_names}{logical skalar (optional), should the species label
be added?}

\item{plot_name_size}{numeric skalar (optional), the species label size}

\item{font_family}{string skalar (optional), the species label font family}

\item{plot}{logical scalar (optional), toggle the output to be either a plot
or a list of plots}
}
\description{
\code{hypo_legend_single} combines several single hamlet legend elements
}
\details{
The function \code{hypo_legend_single} constructs a single hamlet legend
from a vector of hamlet species and a choosen color map (matching in length).

Hamlet species labels can optionally be included.

Available species:
  "aberrans", "atlahua", "castroaguirrei", "chlorurus", "ecosur",
  "floridae", "gemma", "gumigutta", "guttavarius", "indigo",
  "liberte", "maculiferus", "maya", "nigricans", "providencianus",
  "puella", "randallorum", "tan", "unicolor"
}
\examples{
clr <- viridis::inferno(4)

species <- c('unicolor', 'liberte', 'maya', 'castroaguirrei')

hypo_legend_single(species = species, color_map = clr_single,
  circle_color = 'black', plot_names = TRUE)

}
\seealso{
\code{\link{hypo_legend_pair}},
  \code{\link{hypo_legend_flag_single}}
}
