% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypoimg_recolor_svg.R
\name{hypo_read_svg}
\alias{hypo_read_svg}
\title{Read a cairo svg}
\usage{
hypo_read_svg(file_path)
}
\arguments{
\item{file_path}{string skalar (manatory), the cairo svg to import}
}
\description{
\code{hypo_read_svg} imports a cairo svg image file.
}
\details{
This function reads a (cairo) svg file and turns it into
a 'grid object' (grob). This can then be used to annotate
ggplot plots.

To convert a regular svg into a cairo svg use the
grConvert package:

\code{grConvert::convertPicture("image.svg","image.c.svg")}
}
\examples{
svg_file <- system.file("extdata", "logo.c.svg", package = "hypoimg")
svg <- hypo_read_svg(svg_file)

}
