% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypoimg_recolor_svg.R
\name{hypo_recolor_grob_table}
\alias{hypo_recolor_grob_table}
\title{Creates a table svg color variants}
\usage{
hypo_recolor_grob_table(
  svg,
  layer = 1,
  levels,
  colormap,
  x = rep(0.5, length(levels)),
  y = rep(0.5, length(levels)),
  angle = rep(0, length(levels)),
  width = 1,
  height = 1
)
}
\arguments{
\item{svg}{svg object (manatory), the svg object}

\item{layer}{integer skalar (manatory, default = 1), the layer to recolor}

\item{levels}{vector (manatory), the levels of the intended facets}

\item{colormap}{string vector (manatory), the new colors}

\item{x}{numeric vector (optional), x-center of the annotation (length = length(leveles))}

\item{y}{numeric vector (optional), y-center of the annotation (length = length(leveles))}

\item{angle}{numeric scalar (optional), angle of the annotation}

\item{width}{numeric scalar (optional), width of the annotation}

\item{height}{numeric scalar (optional), height of the annotation}
}
\description{
\code{hypo_recolor_grob_table} turns a svg obect into a table of color variants.
}
\details{
This function creates a tibble of colorvariants from a 'raw' svg.
This table can then be fed \code{into geom_hypo_grob()} to annotate
factes of a ggplot with individual color variants of the svg image.
}
\examples{
svg_file <- system.file("extdata", "logo.c.svg", package = "hypoimg")
svg <- hypo_read_svg(svg_file)
svg_tab <- hypo_recolor_grob_table(svg,LETTERS[1:3], c('red', 'green', 'blue'), layer = 1)
ggplot() +
 geom_hypo_grob(data = svg_tab,
  aes(x =x, y = y, grob = grob)) +
  facet_wrap(grp ~ .)

}
\seealso{
\code{\link{geom_hypo_grob}},
}
