% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypoimg_recolor_svg.R
\name{hypo_recolor_svg}
\alias{hypo_recolor_svg}
\title{Recolor a svg layer}
\usage{
hypo_recolor_svg(svg, layer = 1, color = "darkgray")
}
\arguments{
\item{svg}{svg object (manatory), the svg object}

\item{layer}{integer skalar (optional), the layer to recolor}

\item{color}{string skalar (optional), the new color}
}
\description{
\code{hypo_recolor_svg} redefines the color of a specific svg layer.
}
\details{
This function redefines the color of a specific svg layer.
The svg needs to be read in by hypo_read_svg.
}
\examples{
svg_file <- system.file("extdata", "logo.c.svg", package = "hypoimg")
svg <- hypo_read_svg(svg_file)
svg_new <- hypo_recolor_svg(svg, color = 'red')
ggplot() + annotation_custom(grob = svg_new)

}
