% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypoimg_save.R
\name{hypo_save}
\alias{hypo_save}
\title{Add metadata to exported plots}
\usage{
hypo_save(filename, plot = last_plot(),
  comment = "exported with hypoimg::hypo_save()", ...)
}
\arguments{
\item{filename}{File name to create on disk.}

\item{plot}{Plot to save, defaults to last plot displayed.}

\item{comment}{string skalar,comment to add to the figure metadata.}

\item{...}{Other arguments passed on to the \code{ggsave()} function.}
}
\description{
\code{hypo_save()} adds a comment to an exported plot.
}
\details{
Sometimes it can be useful to add metadata to an exported plot.
A typical situation would be to add the name of the r-script
producing the figure.
This can potentially save some time when trying to find the
original script later in life.

Note: The function assumes a UNIX platform and depends on exiftool
to be installed and located within the $PATH
}
\examples{
tibble(x = rnorm(20),y= rnorm(20)) \%>\%
  ggplot(aes(x,y))+geom_point()

hypo_save('~/Desktop/test_plot.png',comment = 'Created by script test.R')
hypo_show_metadata('~/Desktop/test_plot.png')

}
\seealso{
\code{\link{hypo_show_metadata}},
}
