% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypoimg_save.R
\name{hypo_show_metadata}
\alias{hypo_show_metadata}
\title{Read metadata from an exported plot}
\usage{
hypo_show_metadata(filename)
}
\arguments{
\item{filename}{File name to create on disk.}
}
\description{
\code{hypo_show_metadata()} reads metadata from an exported plot.
}
\details{
If we previously added metadata to an exported plot, we need
some way of reading those.
The function \code{hypo_show_metadata()} reports any comment added
with \code{hypo_save()}.

Note: The function assumes a UNIX platform and depends on exiftool
to be installed and located within the $PATH.

See https://www.sno.phy.queensu.ca/~phil/exiftool/index.html
}
\examples{
# only works with exiftools installed - commented out for build
# tibble(x = rnorm(20),y= rnorm(20)) \%>\%
#   ggplot(aes(x,y))+geom_point()
#
# hypo_save('~/Desktop/test_plot.png',comment = 'Created by script test.R')
# hypo_show_metadata('~/Desktop/test_plot.png')
}
\seealso{
\code{\link{hypo_save}},
}
