% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypoimg_data.R
\docType{data}
\name{hypo_trait_img}
\alias{hypo_trait_img}
\title{Hamlet trait drawings.}
\format{
A tibble with 11 rows and 3 variables:
\describe{
\item{trait}{string, trait name}
\item{l}{grob ({grid} object), drawings of the traits, cropped to circle}
\item{r}{grob ({grid} object), drawings of the traits, whole}
}
}
\source{
drawing by KH.
}
\usage{
hypo_trait_img
}
\description{
A data set holding trair specific hamlet drawings.
}
\examples{
ggplot() +
  geom_hypo_grob(data = hypo_trait_img,
                 aes(x = .5, y = .5, grob = grob_circle)) +
  facet_wrap(trait ~ .)

ggplot() +
  geom_hypo_grob(data = hypo_trait_img,
                 aes(x = .5, y = .5, grob = grob_whole)) +
 facet_wrap(trait ~ .)
}
\keyword{datasets}
