% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypoimg_basics.R
\docType{package}
\name{hypoimg}
\alias{hypoimg}
\title{hypoimg: Provides Hypoplectrus image annotations}
\description{
The hypoimg package provides illustrations to annotate
plots within studies of the Caribbean hamlets (Hypoplectrus spp).
}
