% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypoimg_generic.R
\docType{data}
\name{clr2}
\alias{clr2}
\title{two colors}
\format{An object of class \code{character} of length 2.}
\usage{
clr2
}
\description{
\code{clr2} is a combination of two colors I like.
}
\details{
A vector containg blue and orange.
}
\keyword{datasets}
