% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypoimg_generic.R
\name{clr_darken}
\alias{clr_darken}
\title{Programmatically darken colors}
\usage{
clr_darken(color, factor = 0.5)
}
\arguments{
\item{factor}{numeric skalar (optional), between 0 and 1, amount of darkening}

\item{col}{string (color, mandatory)}
}
\description{
\code{clr_darken} produces a darker version on given colors.
}
\details{
Takes a vector of colors and darkens them by a given factor.
1 returns the input color, 0 returns black.
}
\seealso{
\code{\link{clr_lighten}}
}
