% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypoimg_generic.R
\name{clr_lighten}
\alias{clr_lighten}
\title{Programmatically lighten colors}
\usage{
clr_lighten(color, factor = 0.2)
}
\arguments{
\item{factor}{numeric skalar (optional), between 0 and 1, amount of lightening}

\item{col}{string (color, mandatory)}
}
\description{
\code{clr_lighten} produces a lighter version on given colors.
}
\details{
Takes a vector of colors and lightens them by a given factor.
1 returns the input color, 0 returns white
}
\seealso{
\code{\link{clr_darken}}
}
